end_to_end/requirements.txt
mysql-connector-python
psycopg2-binary
pyyaml
docker